refactor(file-editor): 拆出 FileEditor 语言扩展、工具栏、标签栏与交互 hook
